reconvert iter to double when used in setting thetac
Former-commit-id: 9af54c5f2e97602d2b04476fa4b3569bfdcc9207 [formerly 5f430e6c59746298016e11e2a4db5024ef503944]
Former-commit-id: 1be82e5ae3ca4cb8a23e61ebd9a32da87818e6b0

+++ Examples/CapitalConstraints/GlobalSolution.m
@@ -38,7 +38,7 @@ function [ V, C, CB, W, kv, alpha, beta, nu, theta ] = GlobalSolution
     fprintf( '\n' );
     Iter = int32( 0 );
     while true
-        thetac = min( theta, Iter / 100 );
+        thetac = min( theta, double( Iter ) / 100 );
         [ Vnew, Cnew, CBnew ] = IterateValueFunction( V, C, CB, W, kv, av, alpha, beta, nu, thetac );
         e1 = max( max( abs( V - Vnew ) ) );
         e2 = max( max( abs( C - Cnew ) ) );
